+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.17.1
 Released 2016-mm-dd
 
+ - Fix selector in rules metadata.
+
 
 ## Version 0.17.0
 Released 2016-09-28